requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==22.10.0
autopep8==2.0.0
binance-connector==1.18.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.1.3
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.4
cycler==0.11.0
dateparser==1.1.4
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==5.1.0
incremental==22.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
llvmlite==0.37.0
lxml==4.6.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mplfinance==0.12.9b7
multidict==6.0.2
multitasking==0.0.9
numba==0.54.1
numpy==1.20.3
packaging==21.3
pandas==1.3.4
Pillow==9.3.0
plotly==5.3.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.10.31
requests==2.26.0
service-identity==21.1.0
six==1.16.0
tenacity==8.0.1
tomli==2.0.1
Twisted==22.10.0
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
ujson==5.5.0
urllib3==1.26.7
websockets==10.4
Werkzeug==2.2.2
wincertstore==0.2
yarl==1.8.1
yfinance==0.1.64
zipp==3.11.0
zope.interface==5.5.2
